promote: feature/dev → dev (EXP-003 cascade + SQL-MIGRATION-001)

+++ backend/src/application/dto/unit_dto.rs
@@ -1,4 +1,5 @@
 use crate::domain::entities::UnitType;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use validator::Validate;
 
@@ -16,8 +17,8 @@ pub struct CreateUnitDto {
     #[validate(range(min = 0.1))]
     pub surface_area: f64,
 
-    #[validate(range(min = 0.1, max = 1000.0))]
-    pub quota: f64,
+    /// Quote-part en millièmes (Decimal exact, range 0.1..=1000 enforced en domain).
+    pub quota: Decimal,
 }
 
 #[derive(Debug, Deserialize, Validate, Clone)]
@@ -31,8 +32,8 @@ pub struct UpdateUnitDto {
     #[validate(range(min = 0.1))]
     pub surface_area: f64,
 
-    #[validate(range(min = 0.1, max = 1000.0))]
-    pub quota: f64,
+    /// Quote-part en millièmes (Decimal exact, range 0.1..=1000 enforced en domain).
+    pub quota: Decimal,
 }
 
 #[derive(Debug, Serialize)]
@@ -43,6 +44,6 @@ pub struct UnitResponseDto {
     pub unit_type: UnitType,
     pub floor: Option<i32>,
     pub surface_area: f64,
-    pub quota: f64,
+    pub quota: Decimal,
     pub owner_id: Option<String>,
 }
